Correction image

+++ backend/controllers/sauces.js
@@ -45,16 +45,22 @@ exports.modifySauce = (req, res, next) => {
       res.status(401).json({message: 'Non autorisé'});
   } else {
       const filename = sauce.imageUrl.split("/images/")[1];
-      fs.unlink(`images/${filename}`, () => {
-      const sauceObject = req.file ? {
-          ...JSON.parse(req.body.sauce),
-          imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
+      if (req.file) {
+        fs.unlink(`images/${filename}`, () => {
+          const sauceObject = {
+              ...JSON.parse(req.body.sauce),
+              imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
+          }
+          Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+            .then(() => res.status(200).json({ message: "Sauce modifiée !" }))
+            .catch((error) => res.status(400).json({ error }));
+        });
+      } else {
+         const sauceObject = {...req.body};
+         Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+            .then(() => res.status(200).json({ message: "Sauce modifiée !" }))
+            .catch((error) => res.status(400).json({ error }));
       }
-      : { ...req.body };
-      Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-          .then(() => res.status(200).json({ message: "Sauce modifiée !" }))
-          .catch((error) => res.status(400).json({ error }));
-      });
     };
   });
 };
